python-backend/requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
sqlalchemy==2.0.23
python-dotenv==1.0.0
fuzzywuzzy==0.18.0
python-levenshtein==0.23.0
psycopg2-binary==2.9.9
pymysql==1.1.0